Moved user vue files, added axios to register
